pyproject.toml
[project]
name = "demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.82.0",
    "opencv-python>=4.11.0.86",
    "opencv-python-headless>=4.11.0.86",
    "pillow>=11.2.1",
    "pyaudio>=0.2.14",
    "pyttsx3>=2.98",
    "speechrecognition>=3.14.3",
    "vosk>=0.3.45",
]
